Bridge native widths and strengthen cross-ABI tests
Translate CK_ULONG values and nested templates across 32-bit, 64-bit, and Windows layouts. Expand live and mock tests, then split large conversion and test modules.

(cherry picked from commit 0b9f9f4e1d06967ad61899c7805a6a62aa15a7b6)

+++ crates/server/src/server/grpc_service/key_ops/kem.rs
@@ -98,7 +98,7 @@ pub(crate) async fn encapsulate_key(
                 ctx_mgr,
                 &ctx_id,
                 virtual_session,
-                CkObjectHandle(key.0),
+                CkObjectHandle(key.0 as u64),
                 is_token,
             )
             .await;
@@ -199,7 +199,7 @@ pub(crate) async fn decapsulate_key(
                 ctx_mgr,
                 &ctx_id,
                 virtual_session,
-                CkObjectHandle(key.0),
+                CkObjectHandle(key.0 as u64),
                 is_token,
             )
             .await;
