+++ crates/server/src/server/grpc_service/key_ops/wrapping.rs
@@ -218,7 +218,7 @@ async fn unwrap_key_impl(
                 ctx_mgr,
                 &ctx_id,
                 virtual_session,
-                CkObjectHandle(object.0),
+                CkObjectHandle(object.0 as u64),
                 is_token,
             )
             .await;
